backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.3
django-appconf==1.0.6
django-cors-headers==4.6.0
django-imagekit==5.0.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pilkit==3.0
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
